Include 'datadog_agent' class in the catalog when using the generic integration (#697)
We encountered an issue when using `datadog_agent::integrations::generic` where the `::datadog_agent::_agent_major_version` variable was undefined. It looks like the class was not including the `datadog_agent` class in the catalog like all other integrations:

This commit simply adds the include, bringing this integration in line with the rest.